TermAssignment/server/requirement.txt
click==8.0.4
Flask==2.0.3
requests
Jinja2==3.1.1
MarkupSafe==2.1.1
boto3
botocore
Pillow == 9.3
flask-cors
